requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.2.0
pydantic-settings>=2.10.0
apscheduler>=3.11.0
psutil>=7.0.0
ping3>=4.0.0
httpx>=0.28.0
streamlit>=1.49.0
python-telegram-bot>=22.0
librouteros>=3.4.1
pytest>=8.4.0
alembic>=1.16.0
